Create kakute f7 v1.5 4.3 nightly build
everything was fine for a day or 2 trying 4.3 newest nightly an today fc just keeps power cycling an never get last tone . unwired everything after finding nothing with meter probes. the went back to 4.2 everything is fine. new to this trying to help out an post /submit this.hopes it gets to the right place /people